Add ConcatList.ps1 script
The script merges the content of txt files into a MergedContent.txt file
Then it uses it as a base to determine any duplicate

The script is written naively because I barely know PowerShell.
The overall process is rather fast with around 30K+ lines
until the duplication search process happens.
There are ways with Streams but I leave it as is.
Waiting 2/3mn to get a result is currently acceptable.